+++ src/main/java/nextstep/ladder/strategy/BooleanGenerator.java
@@ -1,4 +1,4 @@
-package nextstep.ladder;
+package nextstep.ladder.strategy;
 @FunctionalInterface
 public interface BooleanGenerator {
     boolean generate();
