+++ src/test/java/nextstep/ladder/LineTest.java
@@ -3,6 +3,8 @@ package nextstep.ladder;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import java.util.List;
+
+import nextstep.ladder.strategy.RandomGenerator;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
