+++ src/test/java/nextstep/ladder/TrueGenerator.java
@@ -1,5 +1,7 @@
 package nextstep.ladder;
 
+import nextstep.ladder.strategy.BooleanGenerator;
+
 public class TrueGenerator implements BooleanGenerator {
 
     @Override
